requirements.txt
wheel
discord.py>=2.6.4
GitPython>=3.1.46
psycopg2-binary>=2.9.11
requests>=2.32.5
SQLAlchemy>=2.0.45
ring>=0.10.1
python-dateutil>=2.9.0
